Remove unnecessary header and make private SecurityException and Exception (#4701)
* Refs #20817: Remove all_common.h


* Refs #20817: Remove rtps/network/SenderResource.h


* Refs #20817: Remove fastrtps Exception.h and unused header


* Refs #20817: Remove rtps_fwd.h


* Refs #20817: Delete DomainParticipant::get_resource_event


* Refs #20817: Make Exception.h and SecurityException.h private


* Refs #20817: Move implementation of LoggingLevel in cpp


* Refs #20817: change date in Logging Level


* Refs #20817: Add missing headers


* Refs #20817: Uncrustify


---------
