+++ examples/cpp/rtps/AsSocket/TestWriterSocket.h
@@ -24,7 +24,15 @@
 #include <cstdio>
 #include <string>
 
-#include <fastdds/rtps/rtps_fwd.h>
+namespace eprosima {
+namespace fastrtps {
+namespace rtps {
+class RTPSParticipant;
+class WriterHistory;
+class RTPSWriter;
+} // namespace rtps
+} // namespace fastrtps
+} // namespace eprosima
 
 class TestWriterSocket
 {
