+++ examples/cpp/rtps/Persistent/TestWriterPersistent.h
@@ -20,8 +20,17 @@
 #ifndef TESTWRITERPERSISTENT_H_
 #define TESTWRITERPERSISTENT_H_
 
-#include <fastdds/rtps/rtps_fwd.h>
 #include <fastdds/rtps/writer/WriterListener.h>
+
+namespace eprosima {
+namespace fastrtps {
+namespace rtps {
+class RTPSParticipant;
+class WriterHistory;
+class RTPSWriter;
+} // namespace rtps
+} // namespace fastrtps
+} // namespace eprosima
 
 class TestWriterPersistent
 {
