+++ examples/cpp/rtps/Registered/TestReaderRegistered.h
@@ -21,7 +21,17 @@
 #define TESTREADERREGISTERED_H_
 
 #include <fastdds/rtps/reader/ReaderListener.h>
-#include <fastdds/rtps/rtps_fwd.h>
+
+
+namespace eprosima {
+namespace fastrtps {
+namespace rtps {
+class RTPSParticipant;
+class ReaderHistory;
+class RTPSReader;
+} // namespace rtps
+} // namespace fastrtps
+} // namespace eprosima
 
 class TestReaderRegistered
 {
