+++ include/fastdds/dds/core/policy/ParameterTypes.hpp
@@ -21,12 +21,17 @@
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS_PUBLIC
 
+#include <atomic>
 #include <string>
 #include <vector>
 
 #include <fastcdr/cdr/fixed_size_string.hpp>
 
-#include <fastdds/rtps/common/all_common.h>
+#include <fastdds/rtps/common/InstanceHandle.h>
+#include <fastdds/rtps/common/Locator.h>
+#include <fastdds/rtps/common/SampleIdentity.h>
+#include <fastdds/rtps/common/SerializedPayload.h>
+#include <fastdds/rtps/common/Time_t.h>
 #include <fastdds/rtps/common/Token.h>
 
 #if HAVE_SECURITY
