--- include/fastdds/rtps/common/all_common.h
@@ -1,35 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @file all_common.h
- */
-
-#ifndef _FASTDDS_ALL_COMMON_H_
-#define _FASTDDS_ALL_COMMON_H_
-
-#include <fastdds/rtps/common/Types.h>
-#include <fastdds/rtps/common/CDRMessage_t.h>
-#include <fastdds/rtps/common/Guid.h>
-#include <fastdds/rtps/common/InstanceHandle.h>
-#include <fastdds/rtps/common/Locator.h>
-#include <fastdds/rtps/common/SequenceNumber.h>
-#include <fastdds/rtps/common/FragmentNumber.h>
-#include <fastdds/rtps/common/SerializedPayload.h>
-#include <fastdds/rtps/common/Time_t.h>
-#include <fastdds/rtps/common/CacheChange.h>
-#include <fastdds/rtps/common/MatchingInfo.h>
-
-
-#endif /* _FASTDDS_ALL_COMMON_H_ */
