--- include/fastdds/rtps/rtps_fwd.h
@@ -1,40 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @file rtps_fwd.h
- *
- */
-
-#ifndef _FASTDDS_RTPS_FWD_H_
-#define _FASTDDS_RTPS_FWD_H_
-
-namespace eprosima
-{
-namespace fastrtps
-{
-namespace rtps
-{
-class RTPSWriter;
-class RTPSReader;
-class WriterHistory;
-class ReaderHistory;
-class RTPSParticipant;
-}
-}
-}
-
-
-
-#endif /* _FASTDDS_RTPS_FWD_H_ */
