+++ src/cpp/rtps/security/cryptography/CryptoKeyFactory.h
@@ -30,6 +30,8 @@ namespace fastrtps {
 namespace rtps {
 namespace security {
 
+class SecurityException;
+
 class CryptoKeyFactory
 {
 public:
