+++ src/cpp/rtps/security/cryptography/CryptoTransform.h
@@ -29,6 +29,8 @@ namespace fastrtps {
 namespace rtps {
 namespace security {
 
+class SecurityException;
+
 class CryptoTransform
 {
 public:
