+++ src/cpp/security/cryptography/AESGCMGMAC_KeyExchange.cpp
@@ -28,6 +28,7 @@
 #include <fastdds/dds/log/Log.hpp>
 #include <fastdds/rtps/common/BinaryProperty.h>
 #include <fastdds/rtps/common/Token.h>
+#include <rtps/security/exceptions/SecurityException.h>
 
 // Solve error with Win32 macro
 #ifdef WIN32
