+++ test/blackbox/common/RTPSAsSocketReader.hpp
@@ -51,6 +51,7 @@
 
 #include <fastdds/utils/IPLocator.h>
 #include <fastdds/utils/TimedMutex.hpp>
+#include <gtest/gtest.h>
 
 using eprosima::fastrtps::rtps::IPLocator;
